Fix 4 epic-reviewer blockers; split reporter.py below advisory LOC
WOR-207: add model_param_count TEXT column to BenchRun/bench_store; fix
  OllamaDriver.fetch_model_info() to map parameter_size → model_param_count
  and family → model_family (was incorrectly swapped).

WOR-204: add VRAM headroom ineligibility gate to _is_eligible() — configs
  with headroom < OOM_RISK_HEADROOM_GB (0.5 GB) are disqualified with an
  "OOM risk" reason string.

WOR-201: add conditional APC effectiveness labels to print_apc_section():
  ratio > 1.5× → "APC effective", ratio < 1.1× → "no APC benefit".

WOR-203: add print_concurrency_scaling_section() with per-config
  efficiency labels ("serialised" <50%, "scales well" >80%); called from
  print_ranking() after the APC section.

Advisory fix: split reporter.py (745 LOC, RECOMMEND tier) into four
  focused sub-modules (all < 420 LOC) with reporter.py as a re-export
  facade; all existing callers and test imports unchanged.
